some changes

+++ app/src/main/java/com/soullaundry/BillDetails.java
@@ -11,14 +11,14 @@ public class BillDetails implements Serializable
 
 
    public String clothType;
-    public int qty;
+    public Float qty;
     public int price;
 	public String svcType="";
 	public int discount;
 
 
     public BillDetails(){}
-    public BillDetails(String clothType, int price ,int qty,String svc,int discount){
+    public BillDetails(String clothType, int price ,float qty,String svc,int discount){
         this.clothType=clothType;
         this.qty=qty;
         this.price=price;
@@ -33,7 +33,7 @@ public class BillDetails implements Serializable
         return this.clothType;
     }
 
-    public int getQty(){
+    public float getQty(){
         return this.qty;
     }
     public int getPrice(){
