+++ app/src/main/java/com/soullaundry/CallsFragment.java
@@ -76,7 +76,7 @@ ArrayList<Detail> sending=new ArrayList<Detail>();
                                         Detail d=new Detail();
                                         TextView edi = (TextView) row.getChildAt(0);
                                         d.cloth=edi.getText().toString();
-                                        d.count=Integer.parseInt(ediqty.getText().toString());
+                                        d.count=Float.parseFloat(ediqty.getText().toString());
                                         d.rate=Integer.parseInt(rate.getText().toString());
                                         d.discount=0;
